+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Hover and Go To Definition support for Placeholder Segments that use a Mob/Item/Custom Placeholder/Stat Id (for instance, <caster.stat.example> or <placeholder.test>)
 - RandomSpawns File Type Recognition, with Nodes, Completions, Hovers and all the good stuff
 - Mob Options, DropOptions and DisplayOptions Completions
+- File Completions added by disabled plugins will not be shown
   
 #### Commands
 - `MythicScribe.putSelectionInsideInlineMetaskill` command and context menu item to put the selected text inside of an inline metaskill